# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "High-performance REST API \u0026 real-time OBS Studio Browser Source Overlay Backend built with Golang, Gin, GORM, PostgreSQL, and Goose.",
        "title": "Suporter API - OBS Studio Overlay Backend",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporter Support",
            "email": "support@suporter.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/projects/{uuid}/templates": {
            "get": {
                "description": "Retrieve custom alert templates configured for a project.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Templates"
                ],
                "summary": "List custom overlay templates for a project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a custom alert template with HTML, CSS, and dynamic variables (e.g. {{name}}, {{amount}}, {{description}}).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Templates"
                ],
                "summary": "Create custom overlay template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Template Definition",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Template"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{uuid}/templates/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Templates"
                ],
                "summary": "Get custom template details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Template"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Templates"
                ],
                "summary": "Update custom overlay template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Template",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Template"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Templates"
                ],
                "summary": "Delete custom overlay template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{uuid}/templates/{id}/trigger": {
            "post": {
                "description": "Send dynamic key-value payload (e.g. name, amount, description) to pop up on OBS Studio overlay using custom template HTML/CSS.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Templates"
                ],
                "summary": "Trigger live alert using custom overlay template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dynamic Key-Value Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.TriggerCustomAlertRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.CreateTemplateRequest": {
            "type": "object",
            "required": [
                "event_type",
                "html_template",
                "name"
            ],
            "properties": {
                "css_style": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer",
                    "example": 5000
                },
                "event_type": {
                    "type": "string",
                    "example": "donation"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"name\"",
                        "\"amount\"",
                        "\"description\"]"
                    ]
                },
                "html_template": {
                    "type": "string",
                    "example": "\u003ch1\u003e{{name}} donated {{amount}}\u003c/h1\u003e\u003cp\u003e{{description}}\u003c/p\u003e"
                },
                "name": {
                    "type": "string",
                    "example": "New Donation Alert"
                }
            }
        },
        "domain.EventType": {
            "type": "string",
            "enum": [
                "donation",
                "subscriber",
                "follower",
                "cheer",
                "custom"
            ],
            "x-enum-varnames": [
                "EventTypeDonation",
                "EventTypeSubscriber",
                "EventTypeFollower",
                "EventTypeCheer",
                "EventTypeCustom"
            ]
        },
...
        "domain.Template": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "css_style": {
                    "type": "string",
                    "example": ".alert { background: gold; color: black; }"
                },
                "duration": {
                    "type": "integer",
                    "example": 5000
                },
                "event_type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.EventType"
                        }
                    ],
                    "example": "donation"
                },
                "fields": {
                    "type": "string",
                    "example": "[\"name\",\"amount\",\"description\"]"
                },
                "html_template": {
                    "type": "string",
                    "example": "\u003cdiv class='alert'\u003e\u003ch1\u003e{{name}} donated {{amount}}\u003c/h1\u003e\u003cp\u003e{{description}}\u003c/p\u003e\u003c/div\u003e"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "New Donation Alert"
                },
                "project_id": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "domain.TriggerCustomAlertRequest": {
            "type": "object",
            "required": [
                "payload"
            ],
            "properties": {
                "duration": {
                    "type": "integer",
                    "example": 5000
                },
                "payload": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}